Fix VIA CA2 behavior in PET emulation.
It turns out the VIA CA2 signal has a pull-up so if the CA2 is configured
as an input, it will be driven high in a Commodore PET which uses CA2 to
select the lower-case character set.  I came across a program that (probably
mistakenly) programs the CA2 as an input to choose the lower-case character
et and that broke in emulation.